Add type Reader and Read function
A Redis server can only read from a client a multi-bulk request,
anything else will be an error.

The Redis protocol is well defined, where each multi-bulk request will
comply with this format:

* '*' character
* the number of bulks following
* \r\n

Each bulk will follow, with the following format:

* '$'
* the number of bytes of the data
* \r\n
* n-bytes of data
* \r\n

This might eventually be adapted to the Gedis parser.