Fix gedis.readBulk method
Thanks to @gohelper for pointing out that in case of multiple reads the
previous code won't work. Given that a bulk message contains the length
of bytes to be read, there isn't really a need to multiple reads, so the
code was changed in order to report any error with the byte length.

Also I've added a check when parsing a bulk message contains an invalid
EOF.